Yolo Mars Rover model
We changed to the Yolo model, because our own model was not recognizing any patterns in the data. Perhaps it was because of the data, most likely, or else it might have been the mainly AI-written code. Now we have a model that already works and is even able to detect whether there is an obstacle or not, it just has to be made compatible with our camera and Arduino.