Working AI model with YOLO
The AI model works! A camera on the rover transmits it's footage to a receiver, connected to a laptop which hosts the AI. Then the AI analyses the image and if it doesn't identify any objects, it is a FreePath. Otherwise it would be classified as an obstacle. Now, the Arduino, which is still connected to the laptop, receives this class. If it's obstacle, the rover stops driving and if it's FreePath, the rover drives forwards.

This took a very long time, but we're still not there yet. First we want the Arduino to receive the class via wifi or bluetooth or whatever, not a cable, because it disables us to keep a distance from the rover. Second, the rover shouldn't stop when it sees an obstacle, it should turn around.

We're almost there!